TCA-25 Logic of initial was taking out of the viewmodel
