+++ app/src/main/java/ar/scacchipa/twittercloneapp/viewmodel/MainAppViewModel.kt
@@ -4,35 +4,22 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.preference.PreferenceManager
+import ar.scacchipa.twittercloneapp.domain.StartModeController
 import ar.scacchipa.twittercloneapp.ui.Route
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
 
 class MainAppViewModel(app: Application): AndroidViewModel( app ) {
 
     val sharedPreference = PreferenceManager.getDefaultSharedPreferences( app )
     var splashWasShowed = sharedPreference.getBoolean("splashWasShowed", false)
 
-    var currentPage: MutableStateFlow<String> = MutableStateFlow( getInitialScreen() )
+    var currentPage: MutableStateFlow<String> = MutableStateFlow( Route.SplashScreen.route )
 
-    init {
-        if (!splashWasShowed) {
-            viewModelScope.launch {
-                delay(5000)
-                currentPage.value = Route.LoginScreen.route
+    val startModeController = StartModeController(
+        app = app,
+        scope = viewModelScope,
+        splashCallback = { currentPage.value = Route.SplashScreen.route },
+        loginCallback = { currentPage.value = Route.LoginScreen.route}
+    )
 
-                PreferenceManager
-                    .getDefaultSharedPreferences(app)
-                    .edit()
-                    .putBoolean("splashWasShowed", true)
-                    .apply()
-            }
-        }
-    }
-    fun getInitialScreen() =
-        if (splashWasShowed)
-            Route.LoginScreen.route
-        else
-            Route.SplashScreen.route
 }
